Reorganize database model wrappers; change members members/participants
