+++ authorization/attributes/contexts/event.py
@@ -1,5 +1,5 @@
 from ..context import Context
-from authorization.information import Event as WrappedEvent
+from authorization.information.models import Event as WrappedEvent
 
 
 class EventContext(Context, WrappedEvent):
