+++ authorization/attributes/contexts/group.py
@@ -1,5 +1,5 @@
 from ..context import Context
-from authorization.information import Group as WrappedGroup
+from authorization.information.models import Group as WrappedGroup
 
 
 class GroupContext(Context, WrappedGroup):
