+++ authorization/attributes/resources/group.py
@@ -1,5 +1,5 @@
 from ..resource import Resource
-from authorization.information import Group as WrappedGroup
+from authorization.information.models import Group as WrappedGroup
 
 
 class Group(Resource, WrappedGroup):
