+++ authorization/attributes/resources/user.py
@@ -1,5 +1,5 @@
 from ..resource import Resource
-from authorization.information import User as WrappedUser
+from authorization.information.models import User as WrappedUser
 
 
 class User(Resource, WrappedUser):
