+++ authorization/attributes/subjects/group.py
@@ -1,5 +1,5 @@
 from ..subject import Subject
-from authorization.information import Group as WrappedGroup
+from authorization.information.models import Group as WrappedGroup
 
 
 class Group(Subject, WrappedGroup):
