+++ authorization/attributes/subjects/user.py
@@ -1,5 +1,5 @@
 from ..subject import Subject
-from authorization.information import User as WrappedUser
+from authorization.information.models import User as WrappedUser
 
 
 class User(Subject, WrappedUser):
