--- authorization/information/__init__.py
@@ -1,5 +1,0 @@
-from .model import Model
-from .user import User
-from .group import Group
-from .event import Event
-from .invite import UserInvite, GroupInvite, EventInvite
