+++ authorization/information/models/__init__.py
@@ -1,0 +1,3 @@
+from .user import User
+from .group import Group
+from .event import Event
